+++ frontend/src/components/ResizeTable/types.ts
@@ -1,0 +1,6 @@
+import { TableProps } from 'antd';
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export interface ResizeTableProps extends TableProps<any> {
+	onDragColumn?: (fromIndex: number, toIndex: number) => void;
+}
