+++ frontend/src/container/LogsExplorerList/InfinityTableView/types.ts
@@ -1,4 +1,3 @@
-import { ColumnsType } from 'antd/es/table';
 import { UseTableViewProps } from 'components/Logs/TableView/types';
 
 export type InfinityTableProps = {
@@ -6,5 +5,4 @@ export type InfinityTableProps = {
 	infitiyTableProps: {
 		onEndReached: (index: number) => void;
 	};
-	onColumnsChange: (columns: ColumnsType) => void;
 };
